1. 크롤러 모듈화 및 분리 2. 스케줄러 정리
